Refactor api modules
